+++ components/molecule/validationCode/src/validationCode.js
@@ -98,7 +98,7 @@ const MoleculeValidationCode = forwardRef(
             handler(event, {
               value:
                 valueType === 'string'
-                  ? arrayInnerValue.filter(Boolean).join('')
+                  ? arrayInnerValue && arrayInnerValue.filter(Boolean).join('')
                   : arrayInnerValue
             })
           }
@@ -109,9 +109,14 @@ const MoleculeValidationCode = forwardRef(
       <div className={baseClass}>
         <div className={`${baseClass}-header`}>
           <p className={`${baseClass}-header-labelTitle`}>{labelText}</p>
-          <AtomButton onClick={onClearHandler} design={atomButtonDesigns.FLAT}>
-            {deleteButtonTextLabel}
-          </AtomButton>
+          {deleteButtonTextLabel && (
+            <AtomButton
+              onClick={onClearHandler}
+              design={atomButtonDesigns.FLAT}
+            >
+              {deleteButtonTextLabel}
+            </AtomButton>
+          )}
         </div>
         <div className={`${baseClass}-inputContainer`}>
           <div className={`${baseClass}-inputContainer-pinInput`}>