+++ components/atom/button/CHANGELOG.md
@@ -1,5 +1,34 @@
 # CHANGELOG
 
+# 1.77.0 (2021-12-02)
+
+
+### Bug Fixes
+
+* **components/atom/button:** bump ([459a92e](https://github.com/SUI-Components/sui-components/commit/459a92e55b1c5b91f24d702fed70cc01dd36587d))
+* **components/atom/button:** bump ([9f47029](https://github.com/SUI-Components/sui-components/commit/9f47029da9e54a3b7b5095179a9845262b92f86c))
+
+
+### Features
+
+* **components/atom/button:** adds token to button group padding ([7862c92](https://github.com/SUI-Components/sui-components/commit/7862c922045fdc6f6e34f9c7ab840adcc276d3d8))
+* **components/atom/button:** bump ([1a52acb](https://github.com/SUI-Components/sui-components/commit/1a52acb3923618ee32cd18765bf4ac7dafc1d057))
+* **components/atom/button:** bump ([e8408b6](https://github.com/SUI-Components/sui-components/commit/e8408b64050b10febca3faaae4fa80edd26d6367))
+* **components/atom/button:** bump ([3ba1734](https://github.com/SUI-Components/sui-components/commit/3ba17341986ad724fff6517c777aa2a168c7033e))
+* **components/atom/button:** improvements ([3a5736d](https://github.com/SUI-Components/sui-components/commit/3a5736d497b8fdf2fd1abdafab310761d7eacab6))
+
+
+
+# 1.72.0 (2021-11-30)
+
+
+### Features
+
+* **components/atom/buttom:** atom button selected ([274b9d6](https://github.com/SUI-Components/sui-components/commit/274b9d6cab34b890f26e3ff9beb38572006e330a))
+* **components/atom/button:** bump ([0f0992b](https://github.com/SUI-Components/sui-components/commit/0f0992b89200dc986ca5bc11935b988de80fa651))
+
+
+
 # 1.71.0 (2021-11-25)